asasasa

+++ Ah.Model/Entities/Product.cs
@@ -13,7 +13,7 @@ namespace Ah.Model.Entities
         public int ProductID { get; set; }
         public string? ProductName { get; set; }
         public decimal? UnitPrice { get; set; }
-        public int? UnitsInStock { get; set; }
+        public short? UnitsInStock { get; set; }
         public int? CategoryID { get; set; }
         //Navigation Property
         public Category? Category { get; set; }
